pyproject.toml
[project]
name = "loyalty-telegram-bot"
version = "0.1.0"
description = "Telegram bot for loyalty system"
requires-python = ">=3.8"

dependencies = [
    "python-telegram-bot==20.7",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "gunicorn==21.2.0",
    "flask==2.3.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"